requirements.txt
keras==2.15.0
numpy==1.25.2
Pillow==9.2.0
Flask==3.0.0
Werkzeug==3.0.1